requirements.txt
fastapi
uvicorn
openai
pinecone
ffmpeg-python
python-multipart
